requirements.txt

# Core ML
torch>=2.0.0                    # PyTorch backend
transformers>=4.30.0            # For BLIP captions 
sentence-transformers>=2.2.0    # CLIP model 
faiss-cpu>=1.7.4                # Fast similarity search

# Web & API
streamlit>=1.28.0               # Web interface
requests>=2.31.0                # API calls
python-dotenv>=1.0.0            # Environment variables

# Image Processing
Pillow>=10.0.0                  # Image handling
numpy>=1.24.0                   # Array operations

# Utilities
tqdm>=4.65.0                    # Progress bars
ipython>=3.10.0                 # IPython interactive shell
jupyter>=1.0.0                  # Jupyter notebook